🧬 Migrate the data models about settings in section *template* to module *api_config.template*.
